move spin out to a separate function
observation from @swlynch99 was that the asm for lock() was long,
with the spin loop getting unrolled inline. This would affect the
likelihood of lock() getting inlined in real projects. By moving
the contended lock and spin functions out, the lock() code is
shorter than the std mutex, so it should be short enough.